ai/requirements.txt
openai
google-generativeai
Pillow
torch
sentence-transformers
scikit-learn
langchain
faiss-cpu
llama-cpp-python
flask
langchain-community
langchain-text-splitters
sentence-transformers
faiss-cpu
numpy<2.0
langchain-huggingface
langchain-openai
google-genai